+++ Package.swift
@@ -7,7 +7,7 @@ let package = Package(
     name: "RichTextKit",
     platforms: [
         .iOS(.v14),
-        .macOS(.v12),
+        .macOS(.v11),
         .tvOS(.v14),
         .watchOS(.v7)
     ],